Ajout de la partie 1.4 du projet: mutler a été ajouté et configureé, les route post put et delete ont été modifié pour permettre respectivement l'ajout d'image, la modification d'image et la suppression d'image en même temps que les livres.

+++ controllers/Book.js
@@ -1,3 +1,5 @@
+const fs = require('fs');
+
 const Book = require('../models/Book');
 
 exports.getAllBooks = (req, res, next) => {
@@ -21,21 +23,51 @@ exports.getBestRatedBook = (req, res, next) => {
 };
 
 exports.createBook = (req, res, next) => {
-    delete req.body._id;
-    const book = new Book({ ...req.body });
+    const bookObject = JSON.parse(req.body.book);
+    delete bookObject._id;
+    delete bookObject._userId;
+    const book = new Book({
+        ...bookObject,
+        userId: req.auth.userId,
+        imageUrl: `${req.protocol}://${req.get('host')}/images/${req.file.filename}`
+    });
     book.save()
     .then(() => res.status(201).json({ message: 'Livre enregistré !' }))
     .catch(error => res.status(400).json({ error }));
 };
 
 exports.modifyBook = (req, res, next) => {
-    Book.updateOne({ _id: req.params.id }, { ...req.body, _id: req.params.id })
-        .then(() => res.status(200).json({ message: 'Livre modifié !' }))
-        .catch(error => res.status(400).json({ error }));
+    const bookObject = req.file ? {
+        ...JSON.parse(req.body.book),
+        imageUrl: `${req.protocol}://${req.get('host')}/images/${req.file.filename}`
+    } : { ...req.body };
+    delete bookObject._userId;
+    Book.findOne({ _id: req.params.id })
+        .then((book) => {
+            if (book.userId != req.auth.userId) {
+                res.status(401).json({ message: 'Non autorisé' });
+            } else {
+                Book.updateOne({ _id: req.params.id }, { ...bookObject, _id: req.params.id })
+                    .then(() => res.status(200).json({ message: 'Livre modifié !' }))
+                    .catch(error => res.status(400).json({ error }));
+            }
+        })
+        .catch(error => res.status(400).json({ error }));  
 };
 
 exports.deleteBook = (req, res, next) => {
-    Book.deleteOne({ _id: req.params.id })
-        .then(() => res.status(200).json({ message: 'Livre supprimé !' }))
-        .catch(error => res.status(400).json({ error }));
+    Book.findOne({ _id: req.params.id })
+        .then(() => {
+            if (book.userId != req.auth.userId) {
+                res.status(401).json({message: 'Non autorisé'});
+            } else {
+                const filename = book.imageUrl.split('/images/')[1];
+                fs.unlink(`images/${filename}`, () => {
+                    Book.deleteOne({_id: req.params.id})
+                        .then(() => res.status(200).json({message: 'Livre supprimé !'}))
+                        .catch(error => res.status(401).json({ error }));
+                });
+            }
+        })
+        .catch(error => res.status(500).json({ error }));
 };
